[Refactor] Create MainView folder and Separate GameScene
